Updates for macOS 10.15 (Catalina), various improvements
Use an icon that better matches the visual style of Catalina's toolbar icons.
Use a dark icon when macOS is running in dark mode (at compile time, not runtime,
so a recompile is needed when macOS switches light/dark mode).

Open new windows via the "open" CLI tool, passing a working directory to it;
it's more reliable than asking Terminal to run shell script after its window has opened,
and there's no flash of output while the working directory is changed.
(We still need to ask Terminal to run shell script when opening a new tab, though.)

When opening a tab, if Terminal is already running, use "open -a Terminal" to activate it,
instead of telling Terminal to "activate", so only the frontmost window is brought forward,
letting the rest stay where they were in the z-order, rather than yanking them forward too.

Use a leading space for the "cd" command, to avoid storing it in shell history
(if the shell has been configured to match).

Clear Terminal's scrollback buffer when we clear its screen.

+++ build.sh
@@ -4,8 +4,37 @@
 script_name="Open In Terminal.applescript"
 bundle_name="Open In Terminal.app"
 bundle_id="com.apple.ScriptEditor.id.OpenInTerminal"
-version="1.6"
-copyright="Copyright © 2009-2018 Jason Jackson"
+
+function usage() {
+	script_name="`basename "$0"`"
+	echo 'Builds the "Open In Terminal" Finder-toolbar script as an application.'
+	echo 'See README.md for installation instructions.'
+	echo
+	echo "Usage: $script_name [options]"
+	echo
+	echo "By default, a light or dark icon is chosen automatically, based on"
+	echo "macOS's current setting. Pass --light or --dark to override."
+	exit 1
+}
+
+unset dark
+
+for arg in "$@"; do
+	if [[ $arg == "--dark" ]]; then
+		dark=true
+	elif [[ $arg == "--light" ]]; then
+		dark=false
+	else  # anything else, including -h/--help
+		usage
+	fi
+done
+
+if [[ -z $dark ]]; then
+	dark="$(osascript -e 'tell application "System Events" to tell appearance preferences to log dark mode is true' 2>&1)"
+fi
+
+[[ $dark == true ]] && icon="macOS-10-dark" || icon="macOS-10-light"
+
 
 # --- Utilities ---
 
@@ -24,17 +53,28 @@ function absolute_path() {
 	fi
 }
 
+
 # --- Build Logic ---
 
 # run from the path in which the build script resides
-cd -- "`dirname "$0"`"
+cd -- "$(dirname "$0")"
+
+# find some info in the script
+version="$(head -n 5 "$script_name" | grep -Eo "[0-9.]{3,}")"
+copyright="$(head -n 20 "Open In Terminal.applescript" | grep -E "^Copyright")"
+
+if [[ -z $version || -z $copyright ]]; then
+	echo "Unable to determine bundle version and/or copyright, aborting"
+	exit 1
+fi
 
 # remove any existing version of the app bundle, and create a new one
 rm -rf "$bundle_name"
 osacompile -o "$bundle_name" "$script_name"
+echo "Built $bundle_name"
 
 # copy resources into the bundle
-cp icon/OpenInTerminal.icns "$bundle_name/Contents/Resources/droplet.icns"
+cp "icon/$icon.icns" "$bundle_name/Contents/Resources/droplet.icns"
 cp modifier-keys/modifier-keys "$bundle_name/Contents/Resources"
 
 # fix up Info.plist
@@ -44,7 +84,7 @@ defaults write "$info_plist" CFBundleIdentifier "$bundle_id"
 defaults write "$info_plist" CFBundleShortVersionString "$version"
 defaults write "$info_plist" CFBundleVersion "$version"
 defaults write "$info_plist" LSUIElement 1
-defaults write "$info_plist" NSHumanReadableCopyright "$copyright"
+defaults write "$info_plist" NSHumanReadableCopyright "'$copyright'"
 
 plutil -convert xml1 "$info_plist"
 chmod 644 "$info_plist"
